requirements.txt
# PHP Security Agent Dependencies

# HTTP requests for API calls
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Regular expressions (built-in)
# re - built into Python

# Path handling (built-in)  
# pathlib - built into Python

# JSON handling (built-in)
# json - built into Python

# Data classes (built-in)
# dataclasses - built into Python (Python 3.7+)

# Command line argument parsing (built-in)
# argparse - built into Python

# Type hints (built-in)
# typing - built into Python

# Optional: For enhanced pattern matching
# regex>=2023.6.3

# Optional: For better terminal output
# colorama>=0.4.6

# Optional: For progress bars
# tqdm>=4.65.0

# Optional: For advanced text processing
# beautifulsoup4>=4.12.0

# Development dependencies (uncomment for development)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
